Better abstraction for circuitboard Moved node renderers around
