+++ src/graphics/circuitBoard/features/simulationFeature.h
@@ -7,7 +7,7 @@
 
 
 #include "graphics/circuitBoard/components/cabling/cabling.h"
-#include "graphics/circuitBoard/components/nodes/renderer/nodeRenderers.h"
+#include "graphics/circuitBoard/renderer/node/nodeRenderers.h"
 #include "graphics/circuitBoard/renderer/cablingRenderer.h"
 #include "graphics/circuitBoard/components/nodes/nodePins.h"
 #include "graphics/circuitBoard/components/cabling/cabling.h"
