+++ src/graphics/circuitBoard/renderer/node/notNodeRenderer.h
